Adding presented_by (#19)
Changelog
Updates made to test rails app gemfile to clear vulnerability warnings
Added presented_by class method to Rapido::Controller to set the presenter class.
Added implementation of presenter to Rapido::ApiController to use the presenter class as the presenter of choice if provided, but will fall back to the existing to_h if not
Added tests for to presenter to features/api/messages.feature
Added present_resource_collection for presenting in the index action. Currently just uses presenter, but could use separate presenter (summary_presented_by?) in the future.